Sidebar progress, fonts context API created
Upon creating the sidebar, I stumbled onto some styling difficulties, and I am currently trying to understand how CSS is structured in React.

Other than that, I made a context API for the website fonts with their specific sizes, which I am unaware if it was a good idea, or if it wasn't necessary. I will figure it out when I apply the fonts. In case the drilling of the font style doesn't happen, I will correct it and make a  prop with the fonts.