refactor: :art: Refactor component/scoped styling to use CSS Modules
Test out using CSS Modules as a potential better strategy for React component styling, moving from function-based styles within the component.Reason: To avoid naming collisions.

Not currently sure if it's functional. Even with npm intellisense installed, it cannot automatically find the file path when the CSS module is imported in the component.